requirements.txt
imageio==2.9.0
keyboard==0.13.5
pdf2image==1.14.0
Pillow==8.1.0
pyPdf==1.13
PyPDF2==1.26.0
pyttsx3==2.90
